requirements.txt
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.13.1
cssselect2==0.2.2
defusedxml==0.6.0
Django==2.1.7
django-crispy-forms==1.7.2
django-easy-pdf==0.1.1
html5lib==1.0.1
Pillow==5.4.1
pycparser==2.19
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2018.9
reportlab==3.5.31
six==1.12.0
tinycss2==1.0.2
WeasyPrint==50
webencodings==0.5.1
xhtml2pdf==0.2.3
